requirements.txt
# HTTP & async
requests
aiohttp

# PDF parsing
pymupdf

# LLM
groq

# Environment
python-dotenv

# GitHub / web validation
playwright

# NLP / embeddings
sentence-transformers
scikit-learn

# UI
streamlit

# Date parsing (LinkedIn classifier)
python-dateutil

# Bot detection bypass for LinkedIn fetching
playwright-stealth
